Comments and bug fix on percentage of completed requests

+++ src/main/java/inesc/slave/reports/ThreadReport.java
@@ -109,7 +109,8 @@ public class ThreadReport {
         sb.append("\n Transactions: " + nTransactions + "\n");
         sb.append("Success: " + successTransactions + "\n");
         sb.append("Fail: " + failTransactions + "\n");
-        sb.append("Success Rate: " + (successTransactions / nTransactions) * 100 + "% \n");
+        sb.append("Success Rate: " + ((double) successTransactions / nTransactions) * 100
+                + "% \n");
 
         sb.append("Transaction Rate: " + String.format("%.2f", transactionRate)
                 + " req/sec \n\n");
